requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
isort==5.9.3
numpy==1.21.3
pandas==1.3.4
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
urllib3==1.26.7
